Project structure fixes, got rid of Client structure
