+++ RickAndMortyApp/Model/Character.swift
@@ -45,71 +45,6 @@ struct CharacterData: Decodable {
     let created: String
 }
 
-// Struct contains methods to request character information from API and loading/saving
-// it from/to file
-struct CharacterAPIManager {
-    // MARK: - Public
-    init(client: Client) {}
-    
-    func getAllCharacters(completion: @escaping (Result<[CharacterModel], Error>) -> Void) {
-        var allCharacters = [CharacterModel]()
-        NetworkManager.requestByMethod(method: "character") {
-            switch $0 {
-            case .success(let data):
-                if let infoModel: CharacterInfoModel = JSONHandler.decodeJSONData(data: data) {
-                    infoModel.results.forEach {
-                        character in allCharacters.append(CharacterModel(from: character))
-                    }
-                    let charactersDispatchGroup = DispatchGroup()
-                    for index in 2...infoModel.info.pages {
-                        charactersDispatchGroup.enter()
-                        self.getCharactersByPageNumber(pageNumber: index) {
-                            switch $0 {
-                            case .success(let characters):
-                                characters.forEach {
-                                    character in allCharacters.append(CharacterModel(from: character))
-                                }
-                                charactersDispatchGroup.leave()
-                            case .failure(let error):
-                                completion(.failure(error))
-                            }
-                        }
-                    }
-                    charactersDispatchGroup.notify(queue: .main) {
-                        completion(.success(allCharacters))
-                    }
-                }
-            case .failure(let error):
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    func loadCharacterListFromFile(fileName: String, completion: @escaping (Result<CharacterListModel, Error>) -> Void) {
-        FileSystemManager.loadFromFile(fileName: fileName, completion: completion)
-    }
-    
-    func saveCharcterListToFile(characters: [CharacterModel], fileName: String) {
-        let characterList = CharacterListModel(characters: characters)
-        FileSystemManager.saveToFile(data: characterList, fileName: fileName)
-    }
-    
-    // MARK: - Private functions
-    // Gets all (20) characters from page with given number
-    private func getCharactersByPageNumber(pageNumber: Int, completion: @escaping (Result<[CharacterData], Error>) -> Void) {
-        NetworkManager.requestByMethod(method: "character/"+"?page="+String(pageNumber)) {
-            switch $0 {
-            case .success(let data):
-                if let infoModel: CharacterInfoModel = JSONHandler.decodeJSONData(data: data) {
-                    completion(.success(infoModel.results))
-                }
-            case .failure(let error):
-                completion(.failure(error))
-            }
-        }
-    }
-}
-
 // Struct for representing character list response
 struct CharacterInfoModel: Decodable {
     let info: Info
